core code updates
git-svn-id: https://svn.code.sf.net/p/hsqldb/svn/base/trunk@5729 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/HsqlNameManager.java
@@ -69,7 +69,7 @@ public final class HsqlNameManager {
     private static final HsqlNameManager staticManager = new HsqlNameManager();
 
     static {
-        staticManager.serialNumber.set(Integer.MIN_VALUE);
+        staticManager.serialNumber.set(Long.MIN_VALUE);
     }
 
     private static final HsqlName[] autoColumnNames       = new HsqlName[32];
